+++ app/src/main/java/com/example/bookexchange/ViewModels/MakeRequestViewModel.kt
@@ -3,7 +3,6 @@ package com.example.bookexchange.ViewModels
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.bookexchange.AppUtils
 import com.example.bookexchange.Models.Book
 import com.example.bookexchange.Models.Request
 import com.google.firebase.database.FirebaseDatabase
@@ -26,6 +25,7 @@ class MakeRequestViewModel: ViewModel() {
             val first=async { uploadToMe(myBooks,hisBooks,myKey,hisKey) }
             val second=async { uploadToHim(myBooks,hisBooks,myKey,hisKey)  }
 
+
             if(first.await()&&second.await()){
 
                 result.postValue(true);
@@ -42,15 +42,13 @@ class MakeRequestViewModel: ViewModel() {
 
     suspend fun uploadToMe(myBooks:ArrayList<Book>, hisBooks:ArrayList<Book>, myKey:String, hisKey:String): Boolean {
 
-
-
         var  check=true;
         val job=viewModelScope.launch(Dispatchers.IO) {
             val firebaseDatabase = FirebaseDatabase.getInstance().reference
 
 
             val request = Request(myKey, hisKey, true, true, myBooks, hisBooks, "Sent",false,
-                System.currentTimeMillis().toInt() );
+                System.currentTimeMillis().toInt().toLong() );
 
 
             try {
@@ -75,7 +73,7 @@ class MakeRequestViewModel: ViewModel() {
             val firebaseDatabase = FirebaseDatabase.getInstance().reference
 
             val request = Request( hisKey,myKey, false, false,  hisBooks,myBooks, "Received",false,
-                System.currentTimeMillis().toInt() );
+                System.currentTimeMillis().toLong() );
 
             try {
                 firebaseDatabase.child("All Users").child(hisKey).child("Requests").child(hisKey + myKey)
